+++ db-world-backend/src/main/java/com/db/dbworld/dao/user/LoginDataRepository.java
@@ -12,7 +12,7 @@ import java.util.List;
 public interface LoginDataRepository extends JpaRepository<LoginDataEntity, Integer> {
 
     @Query(value = "SELECT COUNT(*) FROM LOGIN_DATA WHERE USER = :userId", nativeQuery = true)
-    Integer totalNumberOfLogin(@Param(value = "userId") Long userId);
+    Long totalNumberOfLogin(@Param(value = "userId") Long userId);
 
     @Query(value = "SELECT * FROM LOGIN_DATA WHERE USER = :userId order by id desc limit 5", nativeQuery = true)
     List<LoginDataEntity> getLoginDataFromUserId (@Param(value = "userId") Long userId);
