+++ db-world-backend/src/main/java/com/db/dbworld/payloads/ResponsePayloads.java
@@ -72,7 +72,7 @@ public class ResponsePayloads {
         private String email;
         private String password;
         private UserDto.UserRole userRole;
-        private int noOfLogin;
+        private Long noOfLogin;
     }
 
     @Getter
